move test under RectorBuilder package
