+++ packages/RectorBuilder/tests/BuilderRector/RectorProvider.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Contrib\Nette\Application\ValidateControlRector;
+namespace Rector\RectorBuilder\Tests\BuilderRector;
 
 use Rector\Contract\Rector\RectorInterface;
 use Rector\RectorBuilder\BuilderRectorFactory;
