+++ packages/RectorBuilder/tests/BuilderRector/config/rector.yml
@@ -1,0 +1,5 @@
+services:
+    _defaults:
+        autowire: true
+
+    Rector\RectorBuilder\Tests\BuilderRector\RectorProvider: ~
