--- tests/Rector/Contrib/Nette/Application/ValidateControlRector/config/rector.yml
@@ -1,5 +1,0 @@
-services:
-    _defaults:
-        autowire: true
-
-    Rector\Tests\Rector\Contrib\Nette\Application\ValidateControlRector\RectorProvider: ~
